+++ src/pages/contract-detail/contract-detail.tsx
